Crate viaduct_reqwest

Structs

ReqwestBackend

Functions

use_reqwest_backend

viaduct_detect_reqwest_backend

A dummy symbol we include so that we can detect whether or not the reqwest backend got compiled in.